refactor: Move `add` command logic into API
In this refactor, we move the logic of the CLI `add` subcommand
into the API class, for better reusability.

We also deprecate the `-f` or `--from-file` argument,
and automatically treat every valid file-path that is not a torrent
nor a metalink as a source file (to read URIs from), recursively.